+++ src/CHANGES.md
@@ -3,3 +3,4 @@
 ## 0.6.1 - 2023-04-01
 * merge LED fix - LED1 shows MqTT state, LED configureable active high/low #839
 * only publish new inverter data #826
+* potential fix of WiFi hostname during boot up #752